+++ messages/hr/views_common_edit.php
@@ -1,12 +1,11 @@
 <?php
-
-return [
-    'Delete' => 'Obriši.',
-    'By setting an url shortcut value, you can create a better readable url for your page. If <b>URL Rewriting</b> is enabled on your site, the value \'mypage\' will result in an url \'www.example.de/p/mypage\'.' => '',
-    'Configuration' => '',
-    'Here you can configure the general settings of your {pageLabel}.' => '',
-    'Inline Editor' => '',
-    'The abstract will be used as stream entry content to promote the actual page. 
+return array (
+  'By setting an url shortcut value, you can create a better readable url for your page. If <b>URL Rewriting</b> is enabled on your site, the value \'mypage\' will result in an url \'www.example.de/p/mypage\'.' => '',
+  'Configuration' => 'Konfiguracija',
+  'Delete' => 'Obriši.',
+  'Here you can configure the general settings of your {pageLabel}.' => '',
+  'Inline Editor' => '',
+  'The abstract will be used as stream entry content to promote the actual page. 
                         If no abstract is given or the page is only visible for admins, no stream entry will be created.' => '',
-    'e.g. http://www.example.de' => '',
-];
+  'e.g. http://www.example.de' => '',
+);