+++ messages/sq/models_Page.php
@@ -1,0 +1,12 @@
+<?php
+
+return [
+    'Abstract' => '',
+    'Navigation' => '',
+    'Only visible for admins' => '',
+    'Open in new window' => '',
+    'Page' => '',
+    'Url shortcut' => '',
+    'View' => '',
+    'page' => '',
+];