+++ messages/sq/modules_template_controller_LayoutAdminController.php
@@ -1,0 +1,5 @@
+<?php
+
+return [
+    'Here you can manage your template layouts. Layouts are the root of your template pages and can not be combined with other templates.' => '',
+];