+++ messages/sq/modules_template_models_CotnainerContent.php
@@ -1,0 +1,7 @@
+<?php
+
+return [
+    'Allow multiple items?' => '',
+    'Allowed Templates' => '',
+    'Render items as inline-blocks within the inline editor?' => '',
+];