+++ messages/sq/views_admin_add.php
@@ -1,0 +1,8 @@
+<?php
+
+return [
+    'Add new {pageType}' => '',
+    'Create new template' => '',
+    'Edit template' => '',
+    'Settings' => '',
+];