+++ modules/template/widgets/views/pageConfigurationButton.php
@@ -1,10 +1,16 @@
 <?php
 
-use yii\helpers\Html;
-use yii\helpers\Url;
-?>
+use humhub\modules\content\components\ContentContainerActiveRecord;
+use humhub\modules\custom_pages\helpers\Url;
 
-<a target="_blank" class="btn btn-danger btn-sm"  href="<?= Url::to(['edit', 'id' => $pageId, 'sguid' => $sguid]) ?>">
+/* @var $pageId integer */
+/* @var $contentContainer ContentContainerActiveRecord */
+/* @var $target string */
+/* @var $btnClass string */
+/* @var $btnStyles string */
+?>
+<a href="<?= Url::toEditPage($pageId, $contentContainer) ?>"<?= empty($target) ? '' : ' target="' . $target. '"'?>
+   class="<?= $btnClass ?>"<?= empty($btnStyles) ? '' : ' style="' . $btnStyles. '"'?>>
     <i class="fa fa-pencil"></i>
     <?= Yii::t('CustomPagesModule.base', 'Edit Page') ?>
 </a>